Fixed a parsing issue and moved parsing code
