+++ src/DataReader/mod.rs
@@ -1,1 +1,5 @@
 pub mod json_reader;
+pub mod parse_edge;
+pub mod parse_invariant;
+pub mod parse_queries;
+pub mod xml_parser;
