+++ src/ModelObjects/mod.rs
@@ -1,9 +1,5 @@
 pub mod component;
-pub mod parse_edge;
-pub mod parse_invariant;
-pub mod parse_queries;
 pub mod queries;
 pub mod representations;
 pub mod statepair;
 pub mod system_declarations;
-pub mod xml_parser;
